+++ cuvarbase/core.py
@@ -1,3 +1,9 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from builtins import range
+from builtins import object
 import numpy as np
 from .utils import gaussian_window, tophat_window, get_autofreqs
 import pycuda.driver as cuda