+++ cuvarbase/pdm.py
@@ -1,3 +1,9 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from builtins import zip
+from builtins import range
 import numpy as np
 import pycuda.driver as cuda
 import pycuda.gpuarray as gpuarray